requirements.txt
transformers
sentence-transformers
torch
pytube
whisper
scikit-learn
python-dotenv
streamlit